feat: Migrate redux directory structure
