+++ src/reducers/index.js
@@ -1,0 +1,10 @@
+// @flow
+import { combineReducers } from 'redux'
+import toDoList from './toDoList'
+
+// eslint-disable-next-line flowtype-errors/show-errors
+const reducers = combineReducers({
+  toDoList
+})
+
+export default reducers
