+++ src/types/toDo.js
@@ -1,0 +1,9 @@
+// @flow
+
+export type ToDoItem = {
+  key: string,
+  title: string,
+  days: number
+}
+
+export type ToDoList = ToDoItem[]
